refactor: rename the 'adapter.in.rest' pathname e classes to 'presentation.controllers'
